chore: split ProjectList into Legacy and New behind newProjectList flag
