+++ frontend/src/component/project/ProjectList/NewProjectList.tsx
@@ -1,0 +1,1 @@
+export const NewProjectList = () => <></>;
